Rewrite CPU
CPU is 32-bit, big-endian.

Reworked some architecture.

Changed memory type from string to vector<uint8_t>, with no typedef.
Memory is binary data, and NULLs are lost in implicit conversions; so
string is unsatisfactory.  Changed references to use `vector<uint8_t>`
instead of typedef so that code that consumes it should be more aware of
the type (in particular the sign).

CharOutputDevice is invoked by a "port", which it reserves at
initialization.  Also replaced iostream with C I/O functions.

Removed Device priority and it's threading.  The devices still need host
threads, but that will come in a later commit.

Added setBios feature to Motherboard.

Added checks to align memory to words.