Implement preemption in basic_os
The kernel constant SCHEDULE_INTERVAL controls the interval that
'schedule' is invoked in the interval timer interrupt handler.

Changed the looping_funcs.s example to express its return of state when
it is rescheduled.

The state is not guaranteed to be correctly restored.  Other tasks can
trample on another task's status register since it is not currently
being committed before invoking interrupt handlers.

Change interrupt invocation to push all registers onto the stack instead
of just the ip.  This allows the interrupt handler to efficiently know
the exact state of the interrupted task.